refactor(storage): split upload_object module
It was getting too big to navigate effectively.
